Delete old migrations
You don't need to keep around your old migration files in a Rails app,
because your database schema should be captured either in schema.rb or
an equivalent SQL file that can be used to regenerate your schema.

Migrations are not the authoritative source for your database schema.
That role falls to either db/schema.rb or an SQL file which Active
Record generates by examining the database. They are not designed to be
edited, they just represent the current state of the database.